Remove ListHandler and ListResponse for currencies, add SearchHandler and SearchValidator
In favour of adopting a consistent Search Handler convention for the application, the ListHandler and ListResponse for currency listing have been removed. Consequently, a new SearchHandler has been created to accommodate the change with the matching SearchValidator added for validation purposes. This update impacts various components and files across the solution and aligns the handling of currencies with the adopted convention.