Update CurrentProductsViewModel and workspace settings
Updated the 'disposeProduct' function in 'CurrentProductsViewModel.kt' to call the 'dispose' function from the ProductRepository. This change allows disposing products via the ProductRepository. The .idea/workspace.xml file was also modified to reflect changes in the development environment including the addition of 'AllProductsScreenPreview - Light' configuration and an update on 'SELECTED_TARGET' in 'ExecutionTargetManager'.