Update UI components and add custom typography
This commit includes several updates to the UI. It introduces a "expires today" tag to product items in the UI. This is accompanied by modifications in the color and theme files to implement custom typography using the Averia Libre font. Also, several strings have been updated and Kotlin breakpoints have been added for debugging purposes.