Update UI components and enhance theming
The commit introduces ListItem for categories, currencies and units. It also updates SearchDialog using OutlinedTextField and Surface from the Material3 library. Additionally, the selection of items in the list dialogues has been improved by making the entire item clickable. Furthermore, the theming capabilities are enhanced with a custom corner size for shapes.