Update UI elements in various screens
The CreateProductScreen now consists a Column for better categorization of form fields. The CurrencyItems display has been refurbished to show symbol and name neatly. In the CurrentProductsScreen, the icon and text alignment in the Floating Action Button has been improved. The dynamicColor in the BiteRightTheme is now default to true. The dismiss button of the DatePicker in ExpirationDateFormField is now an OutlinedButton.