Add profile editing and product deletion features
This commit introduces the ability for users to edit their profile and delete products. Additionally, the product description in the details screen was updated to have a more clear visual display. The edit profile screen is still under development and a 'Coming Soon' banner is shown on accessing this feature for now.