Add API rate limiting and identity header middleware
ASP.NET Core Rate Limiting package was added to project to control the rate of requests that clients can make to an API within a time span. The IdentityIdHeaderMiddleware was also created to append an identity ID header to each request. Further adjustments are made to enable rate limiting based on this identity ID header. This helps in preventing abuse by limiting the number of requests that a client can make, thereby improving security and stability of the application.